Create: utils/ directory for various 'utilities'. To be worked upon later. Move: utils.py into utils/general.py Move: various functions into utils/general.py , distributed.py, kmeans.py and losses.py. Change accordingly the import on all vector_quantizer_pytorch files.
This is a first step towards a full refacto. The goal is to split the gigantic vector_quantize_pytorch.py file into specific files.
