+++ vector_quantize_pytorch/residual_fsq.py
@@ -9,7 +9,7 @@ from torch.cuda.amp import autocast
 from torch.nn import Module
 
 from vector_quantize_pytorch.finite_scalar_quantization import FSQ
-from vector_quantize_pytorch.utils import exists, round_up_multiple
+from vector_quantize_pytorch.utils.general import exists, round_up_multiple
 
 
 def first(list_):
