+++ vector_quantize_pytorch/residual_lfq.py
@@ -11,7 +11,7 @@ from torch.cuda.amp import autocast
 from torch.nn import Module
 
 from vector_quantize_pytorch.lookup_free_quantization import LFQ
-from vector_quantize_pytorch.utils import exists, round_up_multiple
+from vector_quantize_pytorch.utils.general import exists, round_up_multiple
 
 # main class
