--- vector_quantize_pytorch/utils.py
@@ -1,39 +1,0 @@
-from math import ceil
-
-from einops import pack, unpack
-
-
-def pack_one(t, pattern):
-    return pack([t], pattern)
-
-
-def unpack_one(t, ps, pattern):
-    return unpack(t, ps, pattern)[0]
-
-
-def exists(val):
-    return val is not None
-
-
-def default(val, d):
-    return val if exists(val) else d
-
-
-def noop(*args, **kwargs):
-    pass
-
-
-def log(t, eps=1e-5):
-    return t.clamp(min=eps).log()
-
-
-def entropy(prob):
-    return (-prob * log(prob)).sum(dim=-1)
-
-
-def round_up_multiple(num, mult):
-    return ceil(num / mult) * mult
-
-
-def identity(t):
-    return t
